requirements.txt
opencensus-ext-azure==1.1.14
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
astroid==3.3.9
asttokens==3.0.0
async-timeout==4.0.3
azure-functions==1.23.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dep-logic==0.4.11
dill==0.4.0
distlib==0.3.9
docker==7.1.0
dotenv==0.9.9
dulwich==0.22.8
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
flatbuffers==25.2.10
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.39.0
google-cloud==0.34.0
google-genai==1.10.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hishel==0.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
id==1.5.0
idna==3.10
installer==0.7.0
ipykernel==6.29.5
ipython==8.35.0
isort==6.0.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.24
langchain-core==0.3.56
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.39
magika==0.6.1
Markdown==3.8
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
oauth2client==4.1.3
onnxruntime==1.21.1
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pbs-installer==2025.4.9
pdm==2.23.1
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf>=5.26.1,<6.0.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydot==3.0.4
Pygments==2.19.1
pylint==3.3.6
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-call-graph==2.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
requests-toolbelt==1.0.0
resolvelib==1.1.0
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.3
tenacity==9.1.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unearth==0.17.5
urllib3==2.4.0
uv==0.6.14
uvicorn==0.34.2
virtualenv==20.30.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
zstandard==0.23.0
